1. 梳理url和请求参数；2. 加强服务分类/服务模板删除校验
Former-commit-id: 3148819e3834ed2834ee5873e65b4090bb509895

+++ src/source_controller/coreservice/core/process/process_instance_relation.go
@@ -14,6 +14,7 @@ package process
 
 import (
 	"strconv"
+	"strings"
 
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
@@ -175,6 +176,23 @@ func (p *processOperation) DeleteProcessInstanceRelation(ctx core.ContextParams,
 			common.BKDBIN: option.ModuleIDs,
 		}
 	}
+
+	relations := make([]metadata.ProcessInstanceRelation, 0)
+	if err := p.dbProxy.Table(common.BKTableNameProcessInstanceRelation).Find(deleteFilter).All(ctx.Context, &relations); err != nil {
+		blog.Errorf("DeleteProcessInstanceRelation failed, mongodb failed, table: %s, filter: %+v, err: %+v, rid: %s", common.BKTableNameProcessInstanceRelation, deleteFilter, err, ctx.ReqID)
+		return ctx.Error.CCErrorf(common.CCErrCommDBSelectFailed)
+	}
+	templateProcessIDs := make([]string, 0)
+	for _, relation := range relations {
+		if relation.ProcessTemplateID == common.ServiceTemplateIDNotSet {
+			templateProcessIDs = append(templateProcessIDs, strconv.FormatInt(relation.ProcessID, 10))
+		}
+	}
+	if len(templateProcessIDs) > 0 {
+		blog.Errorf("DeleteProcessInstanceRelation failed, mongodb failed, table: %s, filter: %+v, err: %+v, rid: %s", common.BKTableNameProcessInstanceRelation, deleteFilter, err, ctx.ReqID)
+		return ctx.Error.CCErrorf(common.CCErrCoreServiceShouldNotRemoveProcessCreateByTemplate, strings.Join(templateProcessIDs, ","))
+	}
+
 	if err := p.dbProxy.Table(common.BKTableNameProcessInstanceRelation).Delete(ctx, deleteFilter); nil != err {
 		blog.Errorf("DeleteProcessInstanceRelation failed, mongodb failed, table: %s, filter: %+v, err: %+v, rid: %s", common.BKTableNameProcessInstanceRelation, deleteFilter, err, ctx.ReqID)
 		return ctx.Error.CCErrorf(common.CCErrCommDBDeleteFailed)
